refactor: split lib into modules
